pyproject.toml
[project]
name = "lang-agent"
version = "0.1.0"
description = "LangGraph based multi-agent system"
authors = [
    {name = "Author",email = "author@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langgraph (>=0.4.3,<0.5.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "streamlit (>=1.45.0,<2.0.0)",
    "google-api-python-client (>=2.169.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "huggingface-hub (>=0.31.1,<0.32.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "reportlab (>=4.4.0,<5.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pdf2image (>=1.17.0,<2.0.0)"
]

[tool.poetry]
name = "lang-agent"
version = "0.1.0"
description = "AI 검색 에이전트 - 웹 검색 및 문서 검색을 통합한 하이브리드 시스템"
authors = ["Author <author@example.com>"]
readme = "README.md"
packages = [{include = "ai_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.25"
langgraph = "^0.4.3"
langchain-openai = "^0.3.16"
langchain-community = "^0.3.23"
faiss-cpu = "^1.11.0"
streamlit = "^1.45.0"
google-api-python-client = "^2.169.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
sentence-transformers = "^4.1.0"
huggingface-hub = "^0.31.1"
pandas = "^2.2.3"
numpy = "^2.2.5"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"
python-dotenv = "^1.1.0"
python-docx = "^1.1.2"
reportlab = "^4.4.0"
pymupdf = "^1.25.5"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pydantic = "^2.6.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
